requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
transformers
Pillow
pyttsx3
openai-whisper
boto3
azure-identity
azure-mgmt-resource
azure-ai-textanalytics
google-cloud-sdk

streamlit
python-dotenv
speechrecognition
pydub
gtts

